move the new files to utils
